Allow changing the layout of a struct defined in a lib without breaking ABI
Struct definitions now add a symbol with the size of the struct, and
allocating a variable of that type allocates space for that size instead
of the static size of the type. This way, when a library defines a
struct which is allocated by a user of the library, the library can be
updated by adding or removing fields in the struct without needing to
rebuild the user, because the size symbol will provide the new size and
the user will allocate the correct size.
Of course if the user directly accesses the fields, removing them or adding
some at the beginning of the struct still breaks the ABI, but when using
instead only accessor functions that problem is not present.